+++ src/main/java/com/bit/lot/flower/auth/common/valueobject/JWTAuthenticationShouldNotFilterAntMatcher.java
@@ -1,0 +1,8 @@
+package com.bit.lot.flower.auth.common.valueobject;
+
+public class JWTAuthenticationShouldNotFilterAntMatcher {
+
+  public static final String LOGIN_ANT = "/login";
+  public static final String EMAIL_ANT = "/emails";
+  public static final String SIGNUP_ANT = "/signup";
+}
